+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking Change**: when access undefined environmental variable, throw ``IllegalAccessError`` instead of ``SystemError``
 - **Breaking Change**: evaluate script within separate module context
 - **Breaking Change**: after runtime error, not discard symbol state
+- **Breaking Change**: when modifying map object during iteration, always throw ``InvalidOperationError``
 - allow ``=>`` in abbreviate type notation of Func type
 - allow ``->`` in arm expression
 - allow statement in for-init